init go solution update
