+++ solutions/go/init/code/go.mod
@@ -1,0 +1,3 @@
+module git.codecrafters.io/go-git
+
+go 1.16
