+++ solutions/go/init/code/your_git.sh
@@ -6,6 +6,10 @@
 #
 # DON'T EDIT THIS!
 set -e
+
 tmpFile=$(mktemp)
-go build -o "$tmpFile" $(dirname "$0")/app/*.go
-exec "$tmpFile" "$@"
+trap 'rm -f -- "$tmpFile"' EXIT
+
+( cd $(dirname "$0") && go build -o "$tmpFile" ./cmd/mygit )
+
+"$tmpFile" "$@"
