+++ solutions/go/init/explanation.md
@@ -1,11 +1,10 @@
-The entry point for your Git implementation is in `app/main.go`.
+The entry point for your Git implementation is in `cmd/mygit/main.go`.
 
 Study and uncomment the relevant code: 
 
 ```go
 // Uncomment this block to pass the first stage!
- "io/ioutil"
- "os"
+"os"
 ```
 
 ```go
@@ -13,22 +12,22 @@ Study and uncomment the relevant code:
 
 switch command := os.Args[1]; command {
 case "init":
-    for _, dir := range []string{".git", ".git/objects", ".git/refs"} {
-        if err := os.Mkdir(dir, 0755); err != nil {
-            fmt.Printf("Error creating directory: %s\n", err)
-        }
-    }
+	for _, dir := range []string{".git", ".git/objects", ".git/refs"} {
+		if err := os.Mkdir(dir, 0755); err != nil {
+			fmt.Printf("Error creating directory: %s\n", err)
+		}
+	}
 
-    headFileContents := []byte("ref: refs/heads/master\n")
-    if err := ioutil.WriteFile(".git/HEAD", headFileContents, 0644); err != nil {
-	       fmt.Printf("Error writing file: %s\n", err)
-	   }
+	headFileContents := []byte("ref: refs/heads/master\n")
+	if err := os.WriteFile(".git/HEAD", headFileContents, 0644); err != nil {
+		fmt.Printf("Error writing file: %s\n", err)
+	}
 
-    fmt.Println("Initialized git directory")
+	fmt.Println("Initialized git directory")
 
 default:
-    fmt.Printf("Unknown command %s\n", command)
-    os.Exit(1)
+	fmt.Printf("Unknown command %s\n", command)
+	os.Exit(1)
 }
 ```
